Move Enterprise gateway API rendering behind the extensions boundary
The base gateway API render produces the same objects for both variants. The
Enterprise extension supplies the tigera envoy images, the WAF filter's RBAC and
per-namespace identity, and the l7-log-collector alongside envoy.
