+++ pkg/extensions/gatewayapi.go
@@ -1,0 +1,61 @@
+// Copyright (c) 2026 Tigera, Inc. All rights reserved.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package extensions
+
+import (
+	"context"
+
+	corev1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	"github.com/tigera/operator/pkg/controller"
+	"github.com/tigera/operator/pkg/imageoverride"
+	"github.com/tigera/operator/pkg/render"
+)
+
+// GatewayAPIExtension is the variant's hook into the gateway API controller.
+type GatewayAPIExtension interface {
+	// Images overrides the images the gateway API render resolves.
+	Images() *imageoverride.Overrides
+
+	// ExtendInputs resolves what the modifier needs but cannot read for itself.
+	ExtendInputs(ctx context.Context, ci controller.Inputs) (controller.Inputs, error)
+
+	// Modify layers the variant onto a component the controller rendered.
+	Modify(c render.Component, ri render.Inputs) render.Component
+
+	// GatewayNamespaceObjects returns what a namespace hosting a Gateway needs
+	// beyond the trusted bundle the core operator writes there.
+	GatewayNamespaceObjects(namespace string, pullSecrets []*corev1.Secret) []client.Object
+}
+
+// noopGatewayAPI runs the core operator's behavior unchanged.
+type noopGatewayAPI struct{}
+
+func (noopGatewayAPI) Images() *imageoverride.Overrides {
+	return nil
+}
+
+func (noopGatewayAPI) ExtendInputs(_ context.Context, ci controller.Inputs) (controller.Inputs, error) {
+	return ci, nil
+}
+
+func (noopGatewayAPI) Modify(c render.Component, _ render.Inputs) render.Component {
+	return c
+}
+
+func (noopGatewayAPI) GatewayNamespaceObjects(string, []*corev1.Secret) []client.Object {
+	return nil
+}
